# openapi-v1-draft.yaml
openapi: 3.1.0
info:
  title: MLAir API
  version: 0.1.0-draft
  description: |
    Draft API contract for MLAir v1.
    All resources are tenant/project scoped by design.

    **Model registry:** stages `staging` | `production` | `archived`, **per-version approval**
    (`pending_manual_approval` | `approved` | `rejected`), **promote** (production requires `approved`
    unless `ML_AIR_SKIP_APPROVAL_FOR_PROMOTE=1`), optional **serving slot** assignments
    (`candidate` | `challenger` | `champion` | `canary`), plus optional promote webhooks for downstream systems.

    **Serving slots:** `GET|PUT .../models/{model_id}/serving` are mounted when **`ML_AIR_ENABLE_SERVING_SLOTS_HTTP=1`**
    at API process start (default `0`). **`GET /v1/runtime-config`** → **`features`**: `serving_slots_http` mirrors that mount;
    **`strict_dataset_version_required`** mirrors **`ML_AIR_STRICT_DATASET_VERSION_REQUIRED`** (default `1`);
    **`strict_dataset_version_all_post_runs`** is true only when **`ML_AIR_STRICT_DATASET_VERSION_ALL_POST_RUNS=1`**
    and base strict is on (generic **`POST .../runs`**, gated pipeline run, **`check-readiness`** require a pin even without declared inputs).
    **`readiness_allow_legacy_fallback`** mirrors **`ML_AIR_READINESS_ALLOW_LEGACY_FALLBACK`** (implicit latest-head on dataset readiness when **`dataset_version_id`** is omitted).

    **Runs, readiness, datasets, lineage, pipeline versions, tracking, search, and plugins** sections
...
servers:
...
paths:
...
  /v1/auth/login:
    post:
      tags: [Identity]
      summary: Human login (username + password)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Access and refresh tokens
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: invalid_credential
  /v1/auth/refresh:
    post:
      tags: [Identity]
      summary: Rotate access token from refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: New token pair
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "401":
          description: invalid_refresh_token
  /v1/auth/logout:
    post:
      tags: [Identity]
      summary: Revoke current refresh session
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LogoutRequest"
      responses:
        "204":
          description: Session revoked
  /v1/auth/me:
    get:
      tags: [Identity]
      summary: Current user profile and assignments
      security:
        - BearerAuth: []
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JsonValue"
  /v1/users:
    get:
      tags: [Identity]
      summary: List users (admin)
      security:
        - BearerAuth: []
      responses:
        "200":
          description: User list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JsonValue"
    post:
      tags: [Identity]
      summary: Create user (admin)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserCreateRequest"
      responses:
        "201":
          description: Created user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JsonValue"
  /v1/service-accounts:
    get:
      tags: [Identity]
      summary: List service accounts (admin)
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Service account list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JsonValue"
    post:
      tags: [Identity]
      summary: Create service account (admin)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceAccountCreateRequest"
      responses:
        "201":
          description: Created service account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JsonValue"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: |
        Identity JWT from `POST /v1/auth/login`, Service Account secret, or legacy static token
        when `ML_AIR_LEGACY_STATIC_TOKENS=1`. Worker lease APIs may also accept `ML_AIR_WORKER_TOKEN`.
...
  schemas:
...
    JsonValue:
      type: object
      description: Free-form JSON object from API (shape varies by endpoint).
      additionalProperties: true
...
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
    LogoutRequest:
      type: object
      properties:
        refresh_token:
          type: string
          nullable: true
    TokenResponse:
      type: object
      required: [access_token, token_type, expires_in, refresh_token]
      properties:
        access_token:
          type: string
        token_type:
          type: string
          enum: [Bearer]
        expires_in:
          type: integer
        refresh_token:
          type: string
    LoginResponse:
      allOf:
        - $ref: "#/components/schemas/TokenResponse"
        - type: object
          properties:
            user:
              $ref: "#/components/schemas/JsonValue"
    UserCreateRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
        state:
          type: string
          default: active
        is_global_admin:
          type: boolean
          default: false
    ServiceAccountCreateRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
...